@@ -21,12 +21,12 @@
     "font-awesome": "^4.7.0",
     "jsvectormap": "^1.6.0",
     "mui-datatables": "^4.3.0",
-    "next": "^15.1.0",
-    "react": "^18",
+    "next": "^15.1.1",
+    "react": "^18.3.1",
     "react-apexcharts": "^1.4.1",
     "react-awesome-reveal": "^4.2.14",
     "react-circular-progressbar": "^2.1.0",
-    "react-dom": "^18",
+    "react-dom": "^18.3.1",
     "react-flatpickr": "^3.10.13",
     "react-hot-toast": "^2.4.1",
     "react-quill": "^2.0.0",
