@@ -1,21 +1,293 @@
-"use client"; // Add this at the top of the file
-import { useState } from "react";
-import { useRouter } from "next/navigation"; // This replaces `useRouter` from 'next/router' in the app directory
-import LogoutButton from "@/components/LogoutButton";
+"use client";
+import { useState, useEffect } from "react";
+import { useRouter } from "next/navigation";
 import React from "react";
-import styles from "./StudentDetails.module.css"; // Import CSS module
 import DefaultLayout from "@/components/Layouts/DefaultLayout";
-import User from "@/components/User/User";
+import MUIDataTable from "mui-datatables";
+import { useGlobalState } from "@/context/GlobalContext";
+import { deleteStudentBluk, fetchStudentData } from "@/services/studentService";
+import styles from "./StudentDetails.module.css"; // Import CSS module
+import Loader from "@/components/common/Loader";
+import {
+  fetchsectionByClassData,
+  fetchsectionData,
+} from "@/services/sectionsService"; // Import your section API service
+import { getClasses } from "@/services/classesService"; // Import your classes API service
+import { ThemeProvider } from "@mui/material/styles";
+import useColorMode from "@/hooks/useColorMode";
+import { darkTheme, lightTheme } from "@/components/theme/theme";
+import {
+  Edit,
+  Delete,
+  Visibility,
+  TextFields,
+  AttachMoney,
+} from "@mui/icons-material";
+import IconButton from "@mui/material/IconButton";
+import {
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle,
+  Button,
+  TextField,
+} from "@mui/material";
+import { toast } from "react-toastify";
+import { useLoginDetails } from "@/store/logoStore";
 
-const StudentAdmission = () => {
+const columns = [
+  "Exam Group",
+  "Exam",
+  "Session",
+  "Class",
+  "Section",
+];
+
+const options = {
+  filterType: "checkbox",
+  serverSide: true,
+  responsive: "standard",
+  search: false,
+  selectableRows: "none", 
+  filter: false, 
+  viewColumns: false,
+};
+
+const StudentDetails = () => {
+  const [selectedRows, setSelectedRows] = useState([]);
+  const [colorMode, setColorMode] = useColorMode();
+  const [data, setData] = useState<Array<Array<string>>>([]);
+  const { themType, setThemType } = useGlobalState(); //
+  const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
+  const [page, setPage] = useState(0);
+  const [rowsPerPage, setRowsPerPage] = useState(10);
+  const [totalCount, setTotalCount] = useState(0);
+  const [classes, setClassessData] = useState<Array<any>>([]);
+  const [section, setSections] = useState<Array<any>>([]);
+  const [selectedClass, setSelectedClass] = useState<string | undefined>(
+    undefined,
+  );
+  const [selectedSection, setSelectedSection] = useState<string | undefined>(
+    undefined,
+  );
+  const [keyword, setKeyword] = useState<string>("");
   const router = useRouter();
+
+  const handleDelete = async () => {
+    try {
+      const selectedData = selectedRows.map((rowIndex) => data[rowIndex]); // Map indices to data
+
+      const idsToDelete = selectedData.map((row) => row[0]);
+
+      console.log(idsToDelete); // Handle response
+
+      if (
+        window.confirm("Are you sure you want to delete the selected items?")
+      ) {
+        try {
+          const response = await deleteStudentBluk(idsToDelete);
+        } catch (error) {
+          console.error("Error deleting data:", error);
+          alert("Failed to delete selected data.");
+        }
+      }
+    } catch (error) {
+      console.error("Error deleting data:", error);
+      alert("Failed to delete selected data.");
+    }
+  };
+
+  const handleRowSelectionChange = (
+    curRowSelected: { dataIndex: number; index: number }[],
+    allRowsSelected: { dataIndex: number; index: number }[],
+    rowsSelected: [],
+  ) => {
+    setSelectedRows(rowsSelected); // Update selected rows
+  };
+  const formatStudentData = (students: any[]) => {
+    return students.map((student: any) => [
+      student.id,
+      student.exam_group || "N/A",
+      student.exam  || "N/A",
+      student.session || "N/A",
+      student.class || "N/A",
+      student.section || "N/A",
+
+    ]);
+  };
+  const [selectedSessionId, setSelectedSessionId] = useState<string | null>(
+    null,
+  );
+
+  const getselectedSessionId = useLoginDetails(
+    (state) => state.selectedSessionId,
+  );
+  useEffect(() => {
+    setSelectedSessionId(getselectedSessionId);
+  }, []);
+  const fetchData = async (
+    selectedClass?: string,
+    selectedSection?: string,
+    keyword?: string,
+  ) => {
+    try {
+      // Pass selectedClass and selectedSection as parameters to filter data
+      if (selectedClass && selectedSection) {
+        const result = await fetchStudentData(
+          0,
+          0,
+          selectedClass,
+          selectedSection,
+          keyword,
+          selectedSessionId,
+          1,
+        );
+        setTotalCount(result.totalCount);
+        const formattedData = formatStudentData(result.data);
+        setData(formattedData);
+        setLoading(false);
+      } else {
+        setData([]);
+        setLoading(false);
+      }
+    } catch (error: any) {
+      setError(error.message);
+      setLoading(false);
+    }
+  };
+
+  const fetchClassesAndSections = async () => {
+    try {
+      const classesResult = await getClasses();
+      setClassessData(classesResult.data);
+
+      // Fetch sections if a class is selected
+      if (selectedClass) {
+        const sectionsResult = await fetchsectionByClassData(selectedClass);
+        setSections(sectionsResult.data);
+      } else {
+        setSections([]);
+      }
+    } catch (error: any) {
+      setError(error.message);
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchClassesAndSections();
+  }, [selectedClass]);
+
+  useEffect(() => {
+    fetchData(selectedClass, selectedSection, keyword);
+  }, [selectedClass, selectedSection, keyword]);
+
+  const handlePageChange = (newPage: number) => {
+    setPage(newPage);
+  };
+
+  const handleRowsPerPageChange = (newRowsPerPage: number) => {
+    setRowsPerPage(newRowsPerPage);
+    setPage(0);
+  };
+
+  const handleKeywordChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setKeyword(event.target.value);
+  };
+
+  const handleSearch = () => {
+    setPage(0);
+    fetchData(selectedClass, selectedSection, keyword);
+  };
+  const handleRefresh = () => {
+    setSelectedClass("");
+    setSelectedSection("");
+    setKeyword("");
+  };
+
+  /* if (loading) return <Loader />; */
+  if (error) return <p>{error}</p>;
 
   return (
     <DefaultLayout>
-      Hello
+      <div className={styles.filters}>
+        <div className={styles.filterGroup}>
+          <label className={styles.label}>
+          Exam Group:
+            <select
+              className={`${styles.select} rounded-lg border-stroke outline-none transition focus:border-primary active:border-primary dark:border-form-strokedark dark:bg-form-input dark:text-white dark:focus:border-primary`}
+            >
+              <option value="">Select</option>
+            </select>
+          </label>
+          <label className={styles.label}>
+          Exam:
+            <select
+              className={`${styles.select} rounded-lg border-stroke outline-none transition focus:border-primary active:border-primary dark:border-form-strokedark dark:bg-form-input dark:text-white dark:focus:border-primary`}
+            >
+              <option value="">Select</option>
+            </select>
+          </label>
+          <label className={styles.label}>
+          Session:
+            <select
+              className={`${styles.select} rounded-lg border-stroke outline-none transition focus:border-primary active:border-primary dark:border-form-strokedark dark:bg-form-input dark:text-white dark:focus:border-primary`}
+            >
+              <option value="">Select</option>
+            </select>
+          </label>
+          <label className={styles.label}>
+          Class:
+            <select
+              className={`${styles.select} rounded-lg border-stroke outline-none transition focus:border-primary active:border-primary dark:border-form-strokedark dark:bg-form-input dark:text-white dark:focus:border-primary`}
+            >
+              <option value="">Select</option>
+            </select>
+          </label>
+          <label className={styles.label}>
+          Section:
+            <select
+              className={`${styles.select} rounded-lg border-stroke outline-none transition focus:border-primary active:border-primary dark:border-form-strokedark dark:bg-form-input dark:text-white dark:focus:border-primary`}
+            >
+              <option value="">Select</option>
+            </select>
+          </label>
+         
+          <div className={styles.searchGroup}>
+            <button onClick={handleSearch} className={styles.searchButton}>
+              Search
+            </button>
+            <button onClick={handleRefresh} className={styles.searchButton}>
+              Reset
+            </button>
+          </div>
+        </div>
+      </div>
+
+      {loading ? (
+        <Loader />
+      ) : (
+        <ThemeProvider theme={themType === "dark" ? darkTheme : lightTheme}>
+          <MUIDataTable
+            title={"Exam Schedule"}
+            data={data}
+            columns={columns}
+            options={{
+              ...options,
+              count: totalCount,
+              page: page,
+              rowsPerPage: rowsPerPage,
+              onChangePage: handlePageChange,
+              onChangeRowsPerPage: handleRowsPerPageChange,
+              onRowSelectionChange: handleRowSelectionChange, // Handle row selection
+              onRowsDelete: handleDelete,
+            }}
+          />
+        </ThemeProvider>
+      )}
     </DefaultLayout>
   );
 };
 
-export default StudentAdmission;
+export default StudentDetails;
